Extract utilization info to a separate package
