+++ cluster-autoscaler/processors/status/scale_down_status_processor.go
@@ -21,6 +21,7 @@ import (
 	"k8s.io/autoscaler/cluster-autoscaler/cloudprovider"
 	"k8s.io/autoscaler/cluster-autoscaler/context"
 	"k8s.io/autoscaler/cluster-autoscaler/simulator"
+	"k8s.io/autoscaler/cluster-autoscaler/simulator/utilization"
 	"k8s.io/autoscaler/cluster-autoscaler/utils/drain"
 	klog "k8s.io/klog/v2"
 )
@@ -35,7 +36,7 @@ type ScaleDownStatus struct {
 }
 
 // SetUnremovableNodesInfo sets the status of nodes that were found to be unremovable.
-func (s *ScaleDownStatus) SetUnremovableNodesInfo(unremovableNodesMap map[string]*simulator.UnremovableNode, nodeUtilizationMap map[string]simulator.UtilizationInfo, cp cloudprovider.CloudProvider) {
+func (s *ScaleDownStatus) SetUnremovableNodesInfo(unremovableNodesMap map[string]*simulator.UnremovableNode, nodeUtilizationMap map[string]utilization.Info, cp cloudprovider.CloudProvider) {
 	s.UnremovableNodes = make([]*UnremovableNode, 0, len(unremovableNodesMap))
 
 	for _, unremovableNode := range unremovableNodesMap {
@@ -45,7 +46,7 @@ func (s *ScaleDownStatus) SetUnremovableNodesInfo(unremovableNodesMap map[string
 			continue
 		}
 
-		var utilInfoPtr *simulator.UtilizationInfo
+		var utilInfoPtr *utilization.Info
 		if utilInfo, found := nodeUtilizationMap[unremovableNode.Node.Name]; found {
 			utilInfoPtr = &utilInfo
 			// It's okay if we don't find the util info, it's not computed for some unremovable nodes that are skipped early in the loop.
@@ -65,7 +66,7 @@ func (s *ScaleDownStatus) SetUnremovableNodesInfo(unremovableNodesMap map[string
 type UnremovableNode struct {
 	Node        *apiv1.Node
 	NodeGroup   cloudprovider.NodeGroup
-	UtilInfo    *simulator.UtilizationInfo
+	UtilInfo    *utilization.Info
 	Reason      simulator.UnremovableReason
 	BlockingPod *drain.BlockingPod
 }
@@ -75,7 +76,7 @@ type ScaleDownNode struct {
 	Node        *apiv1.Node
 	NodeGroup   cloudprovider.NodeGroup
 	EvictedPods []*apiv1.Pod
-	UtilInfo    simulator.UtilizationInfo
+	UtilInfo    utilization.Info
 }
 
 // ScaleDownResult represents the result of scale down.
