+++ cluster-autoscaler/simulator/cluster_test.go
@@ -34,98 +34,6 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestUtilization(t *testing.T) {
-	testTime := time.Date(2020, time.December, 18, 17, 0, 0, 0, time.UTC)
-	gpuLabel := GetGPULabel()
-	pod := BuildTestPod("p1", 100, 200000)
-	pod2 := BuildTestPod("p2", -1, -1)
-
-	nodeInfo := schedulerframework.NewNodeInfo(pod, pod, pod2)
-	node := BuildTestNode("node1", 2000, 2000000)
-	SetNodeReadyState(node, true, time.Time{})
-
-	utilInfo, err := CalculateUtilization(node, nodeInfo, false, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 2.0/10, utilInfo.Utilization, 0.01)
-
-	node2 := BuildTestNode("node1", 2000, -1)
-
-	_, err = CalculateUtilization(node2, nodeInfo, false, false, gpuLabel, testTime)
-	assert.Error(t, err)
-
-	daemonSetPod3 := BuildTestPod("p3", 100, 200000)
-	daemonSetPod3.OwnerReferences = GenerateOwnerReferences("ds", "DaemonSet", "apps/v1", "")
-
-	daemonSetPod4 := BuildTestPod("p4", 100, 200000)
-	daemonSetPod4.OwnerReferences = GenerateOwnerReferences("ds", "CustomDaemonSet", "crd/v1", "")
-	daemonSetPod4.Annotations = map[string]string{"cluster-autoscaler.kubernetes.io/daemonset-pod": "true"}
-
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod, pod2, daemonSetPod3, daemonSetPod4)
-	utilInfo, err = CalculateUtilization(node, nodeInfo, true, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 2.5/10, utilInfo.Utilization, 0.01)
-
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod2, daemonSetPod3)
-	utilInfo, err = CalculateUtilization(node, nodeInfo, false, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 2.0/10, utilInfo.Utilization, 0.01)
-
-	terminatedPod := BuildTestPod("podTerminated", 100, 200000)
-	terminatedPod.DeletionTimestamp = &metav1.Time{Time: testTime.Add(-10 * time.Minute)}
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod, pod2, terminatedPod)
-	utilInfo, err = CalculateUtilization(node, nodeInfo, false, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 2.0/10, utilInfo.Utilization, 0.01)
-
-	mirrorPod := BuildTestPod("p4", 100, 200000)
-	mirrorPod.Annotations = map[string]string{
-		types.ConfigMirrorAnnotationKey: "",
-	}
-
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod, pod2, mirrorPod)
-	utilInfo, err = CalculateUtilization(node, nodeInfo, false, true, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 2.0/9.0, utilInfo.Utilization, 0.01)
-
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod2, mirrorPod)
-	utilInfo, err = CalculateUtilization(node, nodeInfo, false, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 2.0/10, utilInfo.Utilization, 0.01)
-
-	nodeInfo = schedulerframework.NewNodeInfo(pod, mirrorPod, daemonSetPod3)
-	utilInfo, err = CalculateUtilization(node, nodeInfo, true, true, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 1.0/8.0, utilInfo.Utilization, 0.01)
-
-	gpuNode := BuildTestNode("gpu_node", 2000, 2000000)
-	AddGpusToNode(gpuNode, 1)
-	gpuPod := BuildTestPod("gpu_pod", 100, 200000)
-	RequestGpuForPod(gpuPod, 1)
-	TolerateGpuForPod(gpuPod)
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod, gpuPod)
-	utilInfo, err = CalculateUtilization(gpuNode, nodeInfo, false, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.InEpsilon(t, 1/1, utilInfo.Utilization, 0.01)
-
-	// Node with Unready GPU
-	gpuNode = BuildTestNode("gpu_node", 2000, 2000000)
-	AddGpuLabelToNode(gpuNode)
-	nodeInfo = schedulerframework.NewNodeInfo(pod, pod)
-	utilInfo, err = CalculateUtilization(gpuNode, nodeInfo, false, false, gpuLabel, testTime)
-	assert.NoError(t, err)
-	assert.Zero(t, utilInfo.Utilization)
-}
-
-func nodeInfos(nodes []*apiv1.Node) []*schedulerframework.NodeInfo {
-	result := make([]*schedulerframework.NodeInfo, len(nodes))
-	for i, node := range nodes {
-		ni := schedulerframework.NewNodeInfo()
-		ni.SetNode(node)
-		result[i] = ni
-	}
-	return result
-}
-
 func TestFindPlaceAllOk(t *testing.T) {
 	node1 := BuildTestNode("n1", 1000, 2000000)
 	SetNodeReadyState(node1, true, time.Time{})
